ci: run every live-backend integration suite on PRs (#609)
* ci: run every live-backend integration suite on PRs — the gap behind 3 prod bugs

All 21 feature-gated integration-<name> suites now run in ci.yml against real
containers (testcontainers: Postgres/Scylla/Redis/Kafka/OpenSearch/MinIO).
This gap hid the three bugs the Keycloak E2E drill found live on prod
(account INT2 decode, inverted optimistic CAS, auth's missing Kafka env) and,
earlier, six of the staging soak's eighteen findings.

fail-fast off, one shared dep cache across the matrix, vm.max_map_count
raised for the OpenSearch-backed suites, 40-min per-suite ceiling. Public
repo: minutes are free; wall time ≈ the slowest Scylla suite.


* ci,test-support: pin container tags to prod versions; feed the raw postgres suite

First-ever CI execution of the 21 suites failed 8: (a) six Scylla-backed
suites — the testcontainers module's floating default tag drifted past prod's
5.4.0 and the newer CQL parser rejects their migrations (local runs kept
passing on cached old images); pin scylla to 5.4.0 and postgres to 16, both
documented to bump in lockstep with prod. (b) the postgres storage suite
predates test-support containers and expects an external DATABASE_URL — CI
now provides one. test-support's self-test rides the scylla pin.


* test-support,postgres: CQL splitter chokes on ';' in inline comments; temp-table tests must ride one connection

Two more finds from putting the suites in CI:
- migrate.rs stripped only full-line comments then split on ';' — a semicolon
  inside an inline comment truncated the surrounding CREATE TABLE, failing six
  services' suites (chat/comment/post/timeline/profile/moderation) and the
  test-support self-test. Now strips inline '--' outside string literals.
- the postgres error-mapping suite created TEMP tables via the POOL: session-
  scoped tables vs per-statement connection checkout = flaky under parallel CI,
  green locally by reuse luck. Each test now acquires one connection.

Chat suite validated locally against real Scylla (5 scenarios green).


* profile: LWT [applied] must be read untyped + ProfileUpdate needs skip_name_checks

Two dormant profile bugs surfaced by the suites' first CI run — profile
create/update had never been exercised against a live Scylla (prod has no
traffic; the soak seeded posts, not profiles):

- LWT results vary in column count ([applied] alone vs + existing row, and
  5.4 returns the full set even on success); strict (bool,) typing failed
  every create. New lwt_applied() reads column 0 untyped, both LWT sites.
- ProfileUpdate used enforce_order WITHOUT skip_name_checks while its field
  names deliberately diverge from bind markers (version appears twice: SET
  new vs IF expected) — every update failed serialization. Same class as the
  soak's social-graph FollowRow finding; now documented at the attribute.

cache_invalidation green locally against real Scylla 5.4; remaining scenario
flakiness locally is Docker contention — the clean CI matrix adjudicates.
Also: test-support removed from the matrix (no integration feature — the
inventory grep had matched its package description).


* post,comment,test-support,profile-tests: three more finds from the CI matrix

- post/comment: LIMIT bound as i64 against CQL int32 — every paged list
  failed serialization (the soak's social-graph #17, same class, never swept
  to siblings; these paths had no live traffic).
- test-support: migration loaders were non-recursive — moderation's split
  migrations/{postgres,scylla} loaded NOTHING and its suite ran against an
  empty schema (relation "cases" does not exist).
- profile harness: random_handle() took the first 12 hex of a UUID v7 — i.e.
  the unix-millis timestamp — so scenarios starting in the same millisecond
  generated IDENTICAL handles: cross-test LWT conflicts produced "fresh
  handle already taken" and the zero-winner claim race. v4 random tail now;
  all 3 scenarios green locally against real Scylla.


* comment: second i64 LIMIT bind (list_replies) — suite green locally


---------

+++ crates/services/comment/src/infrastructure/persistence/scylla_comment_repository.rs
@@ -404,7 +404,11 @@ impl CommentRepository for ScyllaCommentRepository {
         limit:      i32,
         page_token: Option<&str>,
     ) -> Result<(Vec<CommentSummary>, Option<String>), CommentError> {
-        let limit = limit.clamp(1, 100) as i64;
+        // i32, NOT i64: CQL `LIMIT ?` type-checks against int32 and the i64 bind
+        // fails serialization — the soak's social-graph finding (#17), same class,
+        // never swept here (these list paths had no live traffic until the suites
+        // ran in CI).
+        let limit = limit.clamp(1, 100);
         let token = decode_page_token(page_token)?;
 
         let rows: Vec<CommentFeedRow> = if let Some(ref tok) = token {
@@ -460,7 +464,8 @@ impl CommentRepository for ScyllaCommentRepository {
         limit:      i32,
         page_token: Option<&str>,
     ) -> Result<(Vec<CommentSummary>, Option<String>), CommentError> {
-        let limit  = limit.clamp(1, 100) as i64;
+        // i32 for the CQL int32 LIMIT — same fix as list_top_level above.
+        let limit  = limit.clamp(1, 100);
         let token  = decode_page_token(page_token)?;
         let parent = comment_id.as_uuid();
